+++ packages/react-native/template/android/app/src/main/java/com/helloworld/MainApplication.kt
@@ -3,6 +3,7 @@ package com.helloworld
 import android.app.Application
 import com.facebook.react.PackageList
 import com.facebook.react.ReactApplication
+import com.facebook.react.ReactHost
 import com.facebook.react.ReactNativeHost
 import com.facebook.react.ReactPackage
 import com.facebook.react.common.annotations.UnstableReactNativeAPI
@@ -10,7 +11,6 @@ import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.load
 import com.facebook.react.defaults.DefaultReactHost
 import com.facebook.react.defaults.DefaultReactNativeHost
 import com.facebook.react.flipper.ReactNativeFlipper
-import com.facebook.react.interfaces.ReactHost
 import com.facebook.soloader.SoLoader
 
 @UnstableReactNativeAPI
